requirements.txt
psycopg2
python-dotenv
azure-functions
sendgrid